rename body to Home and move to routes
