+++ src/routes/about.jsx
@@ -1,7 +1,7 @@
 function About () {
     return (
         <p>About</p>
-    )
+    );
 }
 
 export default About;
